+++ R/lst_mdl.R
@@ -1,26 +1,26 @@
 list_of_models <- function(x = list()){
-  vctrs::new_vctr(x, class = "lst_mdl")
+  vctrs::new_vctr(x, class = c("mdl_lst", "lst_mdl"))
 }
 
-type_sum.lst_mdl <- function(x){
+type_sum.mdl_lst <- function(x){
   "model"
 }
 
 #' @export
-format.lst_mdl <- function(x, ...){
+format.mdl_lst <- function(x, ...){
   map_chr(x, function(x) paste0("<", model_sum(x), ">"))
 }
 
 #' @export
-vec_cast.character.lst_mdl <- function(x, to, ...) format(x)
+vec_cast.character.mdl_lst <- function(x, to, ...) format(x)
 
 #' @export
-vec_ptype2.lst_mdl.lst_mdl <- function(x, y, ...){
+vec_ptype2.mdl_lst.mdl_lst <- function(x, y, ...){
   list_of_models()
 }
 
 #' @export
-vec_cast.lst_mdl.lst_mdl <- function(x, to, ...){
+vec_cast.mdl_lst.mdl_lst <- function(x, to, ...){
   if(!identical(class(x), class(to))){
     abort("Cannot combine model lists with different reconciliation strategies.")
   }