+++ R/refit.R
@@ -29,7 +29,7 @@ refit.mdl_df <- function(object, new_data, ...){
 }
 
 #' @export
-refit.lst_mdl <- function(object, new_data, ...){
+refit.mdl_lst <- function(object, new_data, ...){
   attrb <- attributes(object)
   object <- map2(object, new_data, refit, ...)
   attributes(object) <- attrb